backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
alembic==1.12.1
python-dotenv==1.0.0
requests==2.31.0
pydantic==2.5.0
python-multipart==0.0.6
scikit-learn==1.3.2
pandas==2.1.4
numpy==1.25.2
nltk==3.8.1
textblob==0.17.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-decouple==3.8
aiofiles==23.2.1
httpx==0.25.2
pytest==7.4.3
pytest-asyncio==0.21.1
beautifulsoup4==4.12.2
lxml==4.9.3 